fix: Moved new shared icons from public into shared components directory
