+++ ui/components/Shared/_newDeisgn/iconComponents/ArrowRightIcon.tsx
