+++ ui/components/Shared/_newDeisgn/iconComponents/CrossIcon.tsx
