+++ ui/components/Shared/_newDeisgn/iconComponents/QrCodeScannerIcon.tsx
